+++ doc/source/whats_new.rst
@@ -78,6 +78,8 @@ Changelog
 
    - Support IIR filters (butterworth, chebyshev, bessel, etc.) by `Eric Larson`_
 
+   - Read labels from FreeSurfer parcellation by  `Martin Luessi`_
+
 API
 ~~~
    - Added nave parameter to source_induced_power() and source_band_induced_power(), use nave=1 by default (wrong nave was used before).
@@ -87,6 +89,7 @@ API
    - Use raw.time_as_index instead of time_to_index (still works but is deprecated).
 
    - The artifacts module (mne.artifacts) is now merged into mne.preprocessing
+
 
 .. _changes_0_4: